Fixed some control issues after reset
The position of the robot was getting reset. Movement are in relation to where it thinks it is. If this is different from where the controller thinks it is the robot could move in the opisit direction and for much further than expected. Need  to uspade the position to the controllers values in future release.